LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-window-object/apis-for-creating-and-navigating-browsing-contexts-by-name/open-features-is-popup-condition-expected.txt

CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'tests[target].forEach')

Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'tests[target].forEach')
